requirements-train.txt
torch==2.4.0
torchvision==0.19.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
redis>=5.0.0
lightgbm>=4.5.0
optuna>=4.0.0
